requirements.txt
accelerate==0.21.0
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
autofaiss==2.15.8
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
blinker==1.6.2
blis==0.7.10
bs4==0.0.1
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.1
confection==0.1.1
contourpy==1.1.1
cryptography==41.0.2
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.14
datasets==2.14.4
deepspeed==0.10.1
diffusers==0.21.2
dill==0.3.7
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.7.0
embedding-reader==1.5.1
et-xmlfile==1.1.0
evaluate==0.4.0
faiss-cpu==1.7.4
filelock==3.12.2
fire==0.4.0
flasgger==0.9.7.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-RESTful==0.3.10
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gensim==4.3.2
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
grpcio==1.58.0
grpcio-tools==1.58.0
h11==0.14.0
h2==4.1.0
hjson==3.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.16.4
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.292
langcodes==3.3.0
langsmith==0.0.37
lit==16.0.6
loguru==0.7.2
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
mistune==3.0.1
mosestokenizer==1.2.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.1.0
Naked==0.1.32
networkx==3.1
ninja==1.11.1
nltk==3.2.4
numexpr==2.8.6
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
openai==0.28.0
opencv-python==4.8.0.76
openfile==0.0.7
openpyxl==3.1.2
pandas==1.5.3
pathtools==0.1.2
pathy==0.10.2
pdfminer.six==20221105
pdfplumber==0.10.1
peft==0.6.0
Pillow==10.0.0
pipreqs==0.4.13
portalocker==2.7.0
preshed==3.0.8
prettytable==3.9.0
protobuf==4.21.12
psycopg2-binary==2.9.7
py-cpuinfo==9.0.0
pyarrow==12.0.1
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.12
pydantic_core==2.6.1
pyext==0.7
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.1
pypdf==3.16.1
pypdfium2==4.18.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==19.0.2
qdrant-client==1.5.4
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
responses==0.18.0
rpds-py==0.10.3
rwkv==0.8.9
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.2
sentencepiece==0.1.99
sentry-sdk==1.29.2
setproctitle==1.3.2
shellescape==3.8.1
similarities==1.1.2
simplejson==3.19.1
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.20
srsly==2.4.7
sympy==1.12
tenacity==8.2.3
termcolor==2.3.0
text2vec==1.2.9
thinc==8.1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
toolwrapper==2.1.0
torch==2.0.1
tqdm==4.66.0
transformers==4.33.1
triton==2.0.0
tushare==1.2.89
typer==0.9.0
typing-inspect==0.9.0
tzdata==2023.3
uctools==1.3.0
urllib3==1.26.16
wandb==0.15.8
wasabi==1.1.2
websocket-client==0.57.0
Werkzeug==2.3.6
xformers==0.0.21
xxhash==3.3.0
yarg==0.1.9
yarl==1.9.2
zipp==3.16.2
